Should check sites accessibility before using registry.in
Output example:

Sorry!  fozz.in  is not availablePlease click here to view new exciting .IN promotionsNumber of allowed queries exceeded.
Sorry!  fozzy.in  is not availablePlease click here to view new exciting .IN promotionsNumber of allowed queries exceeded.
Sorry!  foff.in  is not availablePlease click here to view new exciting .IN promotionsNumber of allowed queries exceeded.